requirements.txt
crewai>=0.1.0
langchain-google-genai>=0.0.5
newspaper3k>=0.2.8
pandas>=2.0.0
python-dotenv>=1.0.0
lxml>=4.9.0
requests>=2.31.0
yfinance>=0.2.28
beautifulsoup4>=4.12.0

# Market-Rover 2.0 - Web UI & Visualizations
streamlit>=1.30.0
plotly>=5.18.0
matplotlib>=3.8.0
reportlab>=4.0.0

# Development & Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
flake8>=6.1.0
black>=23.0.0
isort>=5.12.0
safety>=2.3.0
